Policy exporting /w signing
This commit implements exporting and signing of policy code. A principal
can be bound to an asymmetric keypair (stored in cleartext JSON files as
of now) by using the policy language syntax for doing so. Whenever an
assertion ends with "exportTo PATH" the contents of the policy claim are
serialized and stored in a file and an ECDSA signature is created using
the private key bound to the principal.

There is a test of this functionality in src/test_export_signatures.rs
which uses test_inputs/exporting.

The functionality for importing policy claims (which will also cause the
signature to be checked) is not yet implemented.